home: Fix build on darwin

+++ home/packages.nix
@@ -1,28 +1,32 @@
 {
   flake.homeModules.packages =
-    { pkgs, lib,... }:
+    { pkgs, lib, ... }:
     let
       # multilib for bintools
       multi-bintools = pkgs.wrapBintoolsWith { bintools = pkgs.binutils-unwrapped-all-targets; };
 
       wrapper = pkgs.writeShellApplication {
-        name="bubble-wrapper";
-        runtimeInputs = [pkgs.bubblewrap];
+        name = "bubble-wrapper";
+        runtimeInputs = [ pkgs.bubblewrap ];
         text = lib.readFile ./wrapping.sh;
       };
-      
-      opencode-wrapped = pkgs.symlinkJoin {
-        name = "opencode-wrapped";
-        paths = [ pkgs.opencode ];
-        nativeBuildInputs = [ pkgs.makeWrapper ];
-        postBuild = ''
-          rm -rf "$out/bin"
-          mkdir "$out/bin"
-          makeWrapper "${lib.getExe wrapper}" "$out/bin/opencode" \
-            --set EXECUTABLE "${lib.getExe pkgs.opencode}" \
-            --set PROJECT_ROOT_FILE flake.nix
-        '';
-      };
+
+      opencode-wrapped =
+        if pkgs.stdenv.hostPlatform.isLinux then
+          pkgs.symlinkJoin {
+            name = "opencode-wrapped";
+            paths = [ pkgs.opencode ];
+            nativeBuildInputs = [ pkgs.makeWrapper ];
+            postBuild = ''
+              rm -rf "$out/bin"
+              mkdir "$out/bin"
+              makeWrapper "${lib.getExe wrapper}" "$out/bin/opencode" \
+                --set EXECUTABLE "${lib.getExe pkgs.opencode}" \
+                --set PROJECT_ROOT_FILE flake.nix
+            '';
+          }
+        else
+          pkgs.opencode;
     in
     {
       config.home.packages =
